--- .env
@@ -1,6 +1,0 @@
-COMPOSE_PROJECT_NAME=monopoly
-DB_NAME=monopoly
-DB_PORT=3306
-DB_USER=root
-DB_PASSWORD=464646
-DB_HOST=127.0.0.1
