--- Makefile
@@ -1,2 +1,0 @@
-run:
-		docker-compose up -d --build
