--- docker-compose.yml
@@ -1,39 +1,0 @@
-version: '3.8'
-services:
-  backend:
-    build: ./monopoly_backend
-    command: >
-      sh -c "python manage.py wait_for_db &&
-            python manage.py migrate &&
-            python manage.py runserver 0.0.0.0:8000"
-    ports:
-      - 8000:8000
-    restart: always
-    container_name: server
-    depends_on:
-      - db
-
-  db:
-    image: mysql
-    restart: always
-    hostname: 127.0.0.1
-    environment:
-      # MYSQL_DATABASE: monopoly
-      # MYSQL_USER: test
-      # MYSQL_PASSWORD: test
-      # DB_HOST: 127.0.0.1
-      # MYSQL_ROOT_PASSWORD: test
-      MYSQL_DATABASE: "${DB_NAME}"
-      # MYSQL_USER: "${DB_USER}"
-      MYSQL_PASSWORD: "${DB_PASSWORD}"
-      DB_HOST: "${DB_HOST}"
-      MYSQL_ROOT_PASSWORD: "${DB_PASSWORD}"
-    ports:
-      - 3306:3306
-    container_name: db
-    volumes:
-      - db_volume:/var/lib/mysql
-    command: --character-set-server=utf8 --collation-server=utf8_general_ci
-
-volumes:
-  db_volume:
