--- frontend/join_room.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="ru">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
-    <title>Присоединиться к комнате</title>
-</head>
-<body>
-    <div class="body-container">
-        <div class="body-container__logo"><img src="img/logo.png"></div>
-        <span class="body-container__title">monopoly</span>
-        <a class="body-container__btn" href="id.html">
-        Ввести ID</a>
-        <a class="body-container__btn" href="#">Отсканировать QR - код</a>
-    </div>
-</body>
-</html>
