+++ id.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="ru">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="style.css">
+    <title>Ввести ID</title>
+</head>
+<body>
+    <div class="body-container">
+        <div class="body-container__logo"><img src="img/logo.png"></div>
+        <span class="body-container__title">monopoly</span>
+        <form class="body-container__form">
+            <input class="body-container__form__text" type="id" placeholder ="iD">
+            <input class="body-container__continue" type="submit" value="Присоединиться">
+            <script src="script.js"></script>
+        </form>
+    </div>
+</body>
+</html>
