--- monopoly_backend/Dockerfile
@@ -1,11 +1,0 @@
-FROM python:3
-
-ENV PYTHONUNBUFFERED 1
-
-WORKDIR /app
-
-COPY . .
-
-EXPOSE 8000
-
-RUN pip install -r requirements.txt
