+++ monopoly_backend/monopoly_backend/settings.py
@@ -39,8 +39,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     # ---------------------------
     "rest_framework",
-    "api",
-    "core"
+    "api"
 ]
 
 MIDDLEWARE = [
@@ -81,10 +80,9 @@ DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.mysql',
         'NAME': "monopoly",
-        # "USER": "root",
+        "USER": "root",
         "PASSWORD": "464646",
-        # "HOST": "127.0.0.1",
-        'HOST': 'host.docker.internal',
+        "HOST": "127.0.0.1",
         "PORT": "3306",
         "OPTIONS": {
             "init_command": "SET sql_mode='STRICT_TRANS_TABLES'"
@@ -131,4 +129,4 @@ STATIC_URL = 'static/'
 # Default primary key field type
 # https://docs.djangoproject.com/en/4.2/ref/settings/#default-auto-field
 
-# DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
+DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
