--- monopoly_backend/requirements.txt
@@ -1,5 +1,0 @@
-Django>=3.0,<4.0
-psycopg2>=2.8
-djangorestframework==3.14.0
-mysqlclient==2.2.0
-django-mysql==3.8.1
