+++ style.css
@@ -1,0 +1,135 @@
+* {
+    margin: 0;
+    padding: 0;
+}
+html,body {
+    height: 100%;
+}
+@font-face {
+    font-family: Alice;
+    src: url("fonts/Alice-Regular.ttf");
+}
+.body-container {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    flex-direction: column;
+    width: 100%;
+    min-height: 100%;
+    margin: 0;
+    padding: 0;
+    background-color: rgba(198, 198, 198, 1);;
+}
+.body-container__logo {
+     width: 300px;
+     height: 185px;
+}
+.body-container__logo img {
+    width: 100%;
+    height: 100%;
+}
+.body-container__title {
+    font-family: 'Alice', sans-serif;
+    font-style: normal;
+    font-weight: 400;
+    font-size: 48px;
+    line-height: 55px;
+    text-transform: uppercase;
+    color: #000000;
+    margin-bottom: 50px;
+}
+.body-container__btn {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 284px;
+    height: 35px;
+    background: #AFAFAF;
+    border: 1px solid rgba(0, 0, 0, 0.75);
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    border-radius: 7px;
+    margin-bottom: 20px;
+    font-family: 'Alice', sans-serif;
+    font-style: normal;
+    font-weight: 400;
+    font-size: 20px;
+    line-height: 23px;
+    text-decoration: none;
+    color: #000000;
+}
+.body-container__btn:last-of-type {
+    margin-bottom: 0;
+}
+.body-container__form {
+    display: flex;
+    flex-direction: column;
+}
+.body-container__form__text {
+    width: 284px;
+    height: 35px;
+    left: 41px;
+    top: 487px;
+    background: #D6D6D6;
+    border: 1px solid rgba(0, 0, 0, 0.75);
+    border-radius: 7px;
+    margin-bottom: 40px;
+    box-sizing: border-box;
+    padding: 5px 10px;
+    font-family: 'Alice', sans-serif;
+    font-size: 20px;
+}
+.body-container__continue {
+    width: 180px;
+    height: 32px;
+    margin: 0 auto;
+    margin-bottom: 20px;
+    font-family: 'Alice', sans-serif;
+    font-size: 20px;
+    background-color: #AFAFAF;
+    border: 1px solid rgba(0, 0, 0, 0.75);
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    border-radius: 7px;
+    text-decoration: none;
+    color: #000000;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+.body-container__qr {
+    display: block;
+    min-width: 235px;
+    min-height: 235px;
+    background: #D9D9D9;
+    border: 1px solid rgba(0, 0, 0, 0.75);
+    border-radius: 7px;
+    margin-bottom: 25px;
+}
+.body-container__text {
+    width: 280px;
+    font-family: 'Alice';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 20px;
+    line-height: 23px;
+    text-align: center;
+    margin-bottom: 65px;
+}
+.body-container.top {
+    padding-top: 20px;
+    justify-content: flex-start;
+}
+.body-container__id {
+    box-sizing: border-box;
+    width: 104px;
+    height: 30px;
+    background: #D9D9D9;
+    border: 1px solid rgba(0, 0, 0, 0.75);
+    border-radius: 7px;
+    margin-bottom: 25px;
+    font-family: 'Alice';
+    font-style: normal;
+    font-weight: 400;
+    font-size: 25px;
+    line-height: 29px;
+    padding: 0px 10px;
+}
